Week-05-Dashboards/dash/requirements.txt
dash
dash-bootstrap-components
dash-core-components
dash-html-components
numpy
pandas
plotly-express
seaborn